added separate dtos file in genesis package
